web/requirements.txt
Flask==1.1.1
Flask-Session==0.3.1 
Jinja2==2.10.3
Werkzeug==0.16.0
PyJWT==1.7.1
requests==2.22.0
python-dotenv==0.12.0
pyodbc==4.0.30
pandas==1.0.1
xlrd==1.2.0
openpyxl==3.0.3
scikit-learn==0.22.2
googletrans==2.4.0

plotly==4.6.0
numpy==1.18.2
pandas-datareader==0.8.1